requirements.txt
python-miio
fastapi
uvicorn[standard]